fix: matching incorrect version format.
When we are matching dependencies, if user set the version number as 1-1, it should not be considered as the proper format.
Thus, no match should be found.

Since previously we only use the following expression to find all three components

    const val NAME_EQUAL_STRING_VALUE_REGEX = "([\\w.-]+)\\s*=\\s*\"([^\"]*)\""

this cannot identify whether the version format is correct or not. So a new regex is added :

    const val VERSION_REGEX = "version\\s*=\\s*\\\"([\\d.]+)\\\""

to find the version properly.